+++ run.py
@@ -68,7 +68,7 @@ from voicevox_engine.setting import (
     SettingLoader,
 )
 from voicevox_engine.tts_pipeline import (
-    TTSEngineBase,
+    TTSEngine,
     make_cores,
     make_tts_engines_from_cores,
 )
@@ -145,7 +145,7 @@ def set_output_log_utf8() -> None:
 
 
 def generate_app(
-    synthesis_engines: Dict[str, TTSEngineBase],
+    synthesis_engines: Dict[str, TTSEngine],
     cores: Dict[str, CoreAdapter],
     latest_core_version: str,
     setting_loader: SettingLoader,
@@ -248,7 +248,7 @@ def generate_app(
     def apply_user_dict():
         update_dict()
 
-    def get_engine(core_version: Optional[str]) -> TTSEngineBase:
+    def get_engine(core_version: Optional[str]) -> TTSEngine:
         if core_version is None:
             return synthesis_engines[latest_core_version]
         if core_version in synthesis_engines:
