+++ voicevox_engine/tts_pipeline/__init__.py
@@ -1,7 +1,6 @@
 from ..core_wrapper import CoreWrapper, load_runtime_lib
 from .make_tts_engines import make_cores
 from .tts_engine import TTSEngine, make_tts_engines_from_cores
-from .tts_engine_base import TTSEngineBase
 
 __all__ = [
     "CoreWrapper",
@@ -9,5 +8,4 @@ __all__ = [
     "make_cores",
     "make_tts_engines_from_cores",
     "TTSEngine",
-    "TTSEngineBase",
 ]
